+++ NEWS.md
@@ -1,6 +1,8 @@
 # mirai 0.12.0.9000 (development)
 
-* Fixes `daemons()` specifying 'output = FALSE' being taken as TRUE.
+* Dispatcher sync (and status) timeouts widened to 10s to allow for launching large numbers of daemons.
+* Default for `ssh_config()` argument 'timeout' widened to 10 (seconds).
+* Fixes `daemons()` specifying 'output = FALSE' registering as TRUE instead.
 * Fixes use of `everywhere()` specifying '.args' as an unnamed list or '.expr' as a language object.
 
 # mirai 0.12.0